Progress: setting up the extra collection and model views and templates.
The approach will be to make the user-reading-goal template be the
slider, and the user-reading-goals template be a simple wrapper. Then,
the respective views will do the setup and delegate the slideStop
event mentioned [here](http://www.eyecon.ro/bootstrap-slider/) for
each user-reading-goal.

If there doesn't already exist UserReadingGoals in localStorage, the
slider view will be shown after they're created and passed in to the
user-reading-goals view.

If there does exist UserReadingGoals in localStorage, they'll be
fetched and passed to the user-reading-goals view.

Overall goal is to make sliders that are forms which edit information
stored in localStorage, using BootstrapSlider's events to detect
change, and the save button to detect a save action.